Make JavaScriptCore build with clang++.
Reviewed by Sam Weinig.

* jit/JITInlineMethods.h:
(JSC::JIT::emitPutVirtualRegister):
Explicitly cast to an int.

* yarr/RegexCompiler.cpp:
(JSC::Yarr::compileRegex):
Return 0 instead of false.



Canonical link: https://commits.webkit.org/52213@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@61203 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ JavaScriptCore/jit/JITInlineMethods.h
@@ -645,7 +645,7 @@ ALWAYS_INLINE bool JIT::isOperandConstantImmediateInt(unsigned src)
 ALWAYS_INLINE void JIT::emitPutVirtualRegister(unsigned dst, RegisterID from)
 {
     storePtr(from, Address(callFrameRegister, dst * sizeof(Register)));
-    m_lastResultBytecodeRegister = (from == cachedResultRegister) ? dst : std::numeric_limits<int>::max();
+    m_lastResultBytecodeRegister = (from == cachedResultRegister) ? static_cast<int>(dst) : std::numeric_limits<int>::max();
 }
 
 ALWAYS_INLINE void JIT::emitInitRegister(unsigned dst)
